feat(aci): Make all Detector APIs org-based
